age/requirements.txt
Django>=3.0,<4.0
requests
djangorestframework
librosa==0.8.1
tensorflow